% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwt.R
\name{claim}
\alias{claim}
\title{Generate claim}
\usage{
claim(iss = NULL, sub = NULL, aud = NULL, exp = NULL, nbf = NULL,
  iat = unclass(Sys.time()), jti = NULL, ...)
}
\arguments{
\item{iss}{(Issuer) Claim}

\item{sub}{(Subject) Claim}

\item{aud}{(Audience) Claim}

\item{exp}{(Expiration Time) Claim}

\item{nbf}{(Not Before) Claim}

\item{iat}{(Issued At) Claim}

\item{jti}{(JWT ID) Claim}

\item{...}{additional custom claims to include}
}
\description{
Helper function to create a named list used as the claim of a JWT payload.
See \url{https://tools.ietf.org/html/rfc7519#section-4.1} for details.
}

