% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwk_read.R, R/jwk_write.R
\name{jwk_read}
\alias{jwk}
\alias{jwk_read}
\alias{jwk_write}
\title{JSON web-keys}
\usage{
jwk_read(file)

jwk_write(x, path = NULL)
}
\arguments{
\item{file}{path to file with key data or literal json string}
}
\description{
Read and write RSA, ECDSA or AES keys as JSON web keys.
}
\examples{
# generate an ecdsa key
library(openssl)
key <- ec_keygen("P-521")
jwk_write(key)
jwk_write(as.list(key)$pubkey)

# Same for RSA
key <- rsa_keygen()
jwk_write(key)
jwk_write(as.list(key)$pubkey)
}

