% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwt.R
\name{jwt_encode_hmac}
\alias{jwt_encode_ec}
\alias{jwt_encode_hmac}
\alias{jwt_encode_rsa}
\title{JSON Web Token}
\usage{
jwt_encode_hmac(payload = list(), secret = "secret", size = 256)

jwt_encode_rsa(payload = list(), key, size = 256)

jwt_encode_ec(payload = list(), key = openssl::my_key(), size = 256)
}
\description{
Sign or verify a JSON web token
}
\examples{
# HMAC signing
mysecret <- "This is super secret"
token <- list(name = "jeroen", session = 123456)
jwt_encode_hmac(token, mysecret)

# RSA encoding
mykey <- openssl::rsa_keygen()
jwt_encode_rsa(token, mykey)

# Same with EC
mykey <- openssl::ec_keygen()
jwt_encode_ec(token, mykey)
}

