% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwt.R
\name{jwt_encode_hmac}
\alias{jwt_decode_ec}
\alias{jwt_decode_hmac}
\alias{jwt_decode_rsa}
\alias{jwt_encode_ec}
\alias{jwt_encode_hmac}
\alias{jwt_encode_rsa}
\title{JSON Web Token}
\usage{
jwt_encode_hmac(payload = list(), secret, size = 256)

jwt_decode_hmac(jwt, secret)

jwt_encode_rsa(payload = list(), key, size = 256)

jwt_decode_rsa(jwt, pubkey)

jwt_encode_ec(payload = list(), key, size = 256)

jwt_decode_ec(jwt, pubkey)
}
\arguments{
\item{payload}{a named list with the fields to encode}

\item{secret}{string or raw vector with a secret passphrase}

\item{size}{bitsize of sha2 signature, i.e. \code{sha256}, \code{sha384} or \code{sha512}.}

\item{jwt}{string containing the JSON Web Token (JWT)}

\item{key}{path or object with RSA or EC private key, see \link[openssl:read_key]{openssl::read_key}.}

\item{pubkey}{path or object with RSA or EC public key, see \link[openssl:read_pubkey]{openssl::read_pubkey}.}
}
\description{
Sign or verify a JSON web token. The \code{jwt_encode_hmac}, \code{jwt_encode_rsa},
and \code{jwt_encode_ec} default to \code{HS256}, \code{RS256}, and \code{ES256}
respectively.
}
\examples{
# HMAC signing
mysecret <- "This is super secret"
token <- list(name = "jeroen", session = 123456)
sig <- jwt_encode_hmac(token, mysecret)
jwt_decode_hmac(sig, mysecret)

# RSA encoding
mykey <- openssl::rsa_keygen()
pubkey <- as.list(mykey)$pubkey
sig <- jwt_encode_rsa(token, mykey)
jwt_decode_rsa(sig, pubkey)

# Same with EC
mykey <- openssl::ec_keygen()
pubkey <- as.list(mykey)$pubkey
sig <- jwt_encode_ec(token, mykey)
jwt_decode_ec(sig, pubkey)
}

